refactor: adicionando o swagger + ajustes

+++ src/main/kotlin/br/com/manafood/manafoodorder/adapter/request/commands/create/OrderProductRequest.kt
@@ -1,9 +1,8 @@
 package br.com.manafood.manafoodorder.adapter.request.commands.create
 
-import java.math.BigDecimal
 import java.util.UUID
 
 data class OrderProductRequest(
     val productId: UUID,
-    val quantity: BigDecimal
+    val quantity: Int
 )
