+++ src/main/kotlin/br/com/manafood/manafoodorder/application/usecase/order/commands/create/CreateOrderProductCommand.kt
@@ -1,9 +1,8 @@
 package br.com.manafood.manafoodorder.application.usecase.order.commands.create
 
-import java.math.BigDecimal
 import java.util.UUID
 
 data class CreateOrderProductCommand(
     val productId: UUID,
-    val quantity: BigDecimal
+    val quantity: Int
 )
